requirements.txt
Pillow>=10.0.0
imagehash>=4.3.1
requests>=2.31.0
PyQt6>=6.4.0

# For documentation
Sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0
myst-parser>=1.0.0

# For system information
psutil>=5.9.0

# For Windows
pywin32>=306

# For all platforms
pyinstaller>=5.12.0
pyinstaller-hooks>=1.0.0

# For testing
pytest>=8.3.4
